Added display package for the CLI
Added a directory called 'display' to the cli tree. Added a Canvas class
which represents a mutable array of characters for rendering to the
screen. Added a BoardRenderer class which takes a Board in its
constructor and determines information necessary for rendering that
Board to a Canvas. Added a MarkRenderer whose purpose is to render
whatever Mark is in a certain position in a Board; this avoids having to
create new Renderers every time a Board position is modified.

Signed-off-by: Dylan Frese <dmfrese@gmail.com>

Signed-off-by: Dylan Frese <dmfrese@gmail.com>