AbstractBoard square-only methods throw TypeErrors
A few methods in the AbstractBoard module can only be used (and only
make sense) if that board is square. To be precise, these were
major_diagonal, minor_diagonal, and size. If these methods had been
called on a non-square board, they would raise RuntimeErrors. This has
been changed to TypeErrors.