Board#initialize errors if given block & levels.
The initializer for the Board class now throws an ArgumentError if
levels is defined when a block has been passed in. This is because the
block is used to initialize the array of subboards if it exists, and
levels is not used at all. Code which passes in both may then be
mistaken about the use of that parameter.