Mark#[] always raises a TypeError.
Mark#[] used to return that Mark. However, this can imply that Marks are
infinitely big and infinitely many children, which is the opposite of
what's true. It now always raises a TypeError, which makes more sense
since the index would always be out of bound anyway (either negative or
greater than or equal to the width * height).