Fixed MarkRenderer#render Canvas bug.
MarkRenderer used some previous iteration of the method 'width' and the
method 'height' to create a Canvas if none was specified in the render
method. This has been fixed, and a Canvas of width and height 1 is now
created instead. Additionally, to reduce confusion, the method 'board'
which returned the Mark to be rendered has been renamed to 'Mark'.