Implemented a sane Board#inspect and Mark#inspect
Previously, the Board's inspect method was the default, which included
the parent and all children, which resulted in an infinite loop. The new
inspect method includes the width and height of the board, as well as a
list of all subboards. Mark's inspect implementation simply returns the
string representation of that Mark, since a Mark has no other
information and is uniquely identified by its string representation.