Canvas#[]= aborts if index is out of bounds.
If the index passed to Canvas#[]= is out of bounds, the method simply
returns, doing nothing. This is to prevent the expansion of the @chars
array beyond its originally set size. Here, out of bounds indicates an
index greater than or equal to the size of the array. Negative numbers
are still accepted.