Changed Game#valid? to match Move
The recent change in Move necessitated changes to their validity
checking in the Game class. Previously, valid? checked these three
things:

o Is the Move made by the current player?
o Is the Move made on a descendant of the current board?
o Is the Move made on a space in the bounds of its board?

These are the checks valid? currently makes:

o Is the Move made by the current player?
o Is the Move made on a board that exists?
o Is the Move made on a space that is a Mark?